Moved the outlets and actions of the cpu view into a separate file
